[Http] Moving Client classes to a new Client\Manager namespace for better organization.
