+++ src/Valkyrja/Application/Env.php
@@ -40,8 +40,8 @@ use Valkyrja\Crypt\Manager\SodiumCrypt;
 use Valkyrja\Filesystem\Manager\Contract\Filesystem;
 use Valkyrja\Filesystem\Manager\FlysystemFilesystem;
 use Valkyrja\Filesystem\Manager\LocalFlysystemFilesystem;
-use Valkyrja\Http\Client\Contract\Client;
-use Valkyrja\Http\Client\GuzzleClient;
+use Valkyrja\Http\Client\Manager\Contract\Client;
+use Valkyrja\Http\Client\Manager\GuzzleClient;
 use Valkyrja\Http\Message\Constant\HeaderName;
 use Valkyrja\Http\Message\Enum\SameSite;
 use Valkyrja\Http\Middleware\Contract\RequestReceivedMiddleware as HttpRequestReceivedMiddleware;
