+++ src/Valkyrja/Http/Client/Manager/LogClient.php
@@ -11,11 +11,11 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Client;
+namespace Valkyrja\Http\Client\Manager;
 
 use JsonException;
 use Override;
-use Valkyrja\Http\Client\Contract\Client as Contract;
+use Valkyrja\Http\Client\Manager\Contract\Client as Contract;
 use Valkyrja\Http\Message\Request\Contract\Request;
 use Valkyrja\Http\Message\Response\Contract\Response;
 use Valkyrja\Http\Message\Response\EmptyResponse;
