+++ tests/Unit/Application/ApplicationTest.php
@@ -49,10 +49,10 @@ use Valkyrja\Event\Collection\Contract\Collection as EventCollection;
 use Valkyrja\Event\Collector\Contract\Collector as EventCollector;
 use Valkyrja\Event\Data as EventData;
 use Valkyrja\Event\Dispatcher\Contract\Dispatcher as EventDispatcher;
-use Valkyrja\Http\Client\Contract\Client;
-use Valkyrja\Http\Client\GuzzleClient;
-use Valkyrja\Http\Client\LogClient;
-use Valkyrja\Http\Client\NullClient;
+use Valkyrja\Http\Client\Manager\Contract\Client;
+use Valkyrja\Http\Client\Manager\GuzzleClient;
+use Valkyrja\Http\Client\Manager\LogClient;
+use Valkyrja\Http\Client\Manager\NullClient;
 use Valkyrja\Http\Message\Factory\Contract\ResponseFactory;
 use Valkyrja\Http\Middleware\Cache\CacheResponseMiddleware;
 use Valkyrja\Http\Middleware\Handler\Contract\RequestReceivedHandler;
