+++ tests/Unit/Http/Client/Manager/NullClientTest.php
@@ -11,35 +11,24 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Tests\Unit\Http\Client;
+namespace Valkyrja\Tests\Unit\Http\Client\Manager;
 
-use JsonException;
-use PHPUnit\Framework\MockObject\Exception;
-use Valkyrja\Http\Client\LogClient;
+use Valkyrja\Http\Client\Manager\NullClient;
 use Valkyrja\Http\Message\Request\Request;
 use Valkyrja\Http\Message\Response\EmptyResponse;
-use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Tests\Unit\TestCase;
 
 /**
- * Test the LogClient service.
+ * Test the NullClient service.
  *
  * @author Melech Mizrachi
  */
-class LogClientTest extends TestCase
+class NullClientTest extends TestCase
 {
-    /**
-     * @throws Exception
-     * @throws JsonException
-     */
     public function testSendRequest(): void
     {
-        $logger = $this->createMock(Logger::class);
-
-        $client  = new LogClient($logger);
+        $client  = new NullClient();
         $request = new Request();
-
-        $logger->expects(self::once())->method('info');
 
         $response = $client->sendRequest($request);
